GRHB-82: + usePagination hook
