+++ frontend/src/hooks/hooks.ts
@@ -3,6 +3,7 @@ export { useAppForm } from './use-app-form/use-app-form.hook';
 export { useAppSelector } from './use-app-selector/use-app-selector.hook';
 export { useFormControl } from './use-form-control/use-form-control.hook';
 export { useHandleClickOutside } from './use-handle-click-outside/use-handle-click-outside.hook';
+export { usePagination } from './use-pagination/use-pagination.hook';
 export { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 export {
   useLocation,
